fixup! fixup! fixup! Add getAll method to UserDao

+++ wsmaster/che-core-api-user/src/main/java/org/eclipse/che/api/user/server/jpa/JpaUserDao.java
@@ -182,9 +182,9 @@ public class JpaUserDao implements UserDao {
     }
 
     @Transactional
-    protected int getTotalCount() throws ServerException {
+    protected long getTotalCount() throws ServerException {
         try {
-            return managerProvider.get().createNamedQuery("User.getTotalCount", Long.class).getSingleResult().intValue();
+            return managerProvider.get().createNamedQuery("User.getTotalCount", Long.class).getSingleResult();
         } catch (RuntimeException x) {
             throw new ServerException(x.getLocalizedMessage(), x);
         }
